Move common code into pipx_release.py.
